[AC-7643] move resolve program func to baseUserProfile
