+++ web/impact/impact/graphql/types/__init__.py
@@ -1,5 +1,6 @@
 from impact.graphql.types.user_type import UserType
 from impact.graphql.types.program_type import ProgramType
+from impact.graphql.types.base_user_profile_type import BaseUserProfileType
 from impact.graphql.types.startup_type import StartupType
 from impact.graphql.types.industry_type import IndustryType
 from impact.graphql.types.program_family_type import ProgramFamilyType
