+++ web/impact/impact/tests/test_impact_email_backend.py
@@ -5,10 +5,10 @@ from impact.minimal_email_handler import MinimalEmailHandler
 
 
 IMPACT_BACKEND_PATH = 'impact.impact_email_backend.ImpactEmailBackend'
-# Note that this definition, which duplicates a value in settings.py, 
+# Note that this definition, which duplicates a value in settings.py,
 # is necessary since the value is not included in the settings for test
-# See AC-7670 for a ticket aimed at resolving this and other issues 
-# around testing email backends   
+# See AC-7670 for a ticket aimed at resolving this and other issues
+# around testing email backends
 
 ADD_LOGGING_HEADERS = ".".join(["impact",
                                 "impact_email_backend",
